consumer/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "faststream[cli,rabbit]>=0.6.7",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "tenacity>=9.1.4",
]
